+++ Assets/PrefabsDK/Player.prefab
@@ -31,7 +31,7 @@ Transform:
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 5.31, y: -0.4, z: 0}
-  m_LocalScale: {x: 0.5345086, y: 0.41668317, z: 0.752}
+  m_LocalScale: {x: 0.6205, y: 0.5541, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 0}
