--- Assets/PrefabsDK/SlowingZone.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 8d6f33c02467e4b42a137db9b702fee3
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
