--- Assets/ScriptsDK/GoalMoverTriggerDK.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 5e9564e05dabd7740b121096499bbc48
