--- Assets/ScriptsDK/PlayerObjectInteractionDK.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 93964765d0e403c4a87acf0e0df76188
