--- Assets/ScriptsDK/SlowingZoneDK.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 915258033086aca4cb21cbad7674efdc
